refactor(auction-card): config 파일명을 constants로 변경
