+++ src/entities/auction/ui/auction-item-card/model/constants.ts
