+++ src/entities/auction/ui/auction-item-card/ui/auction-item-card.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 
 import LiveBadge from "@/entities/auction/ui/auction-badge/live-badge";
 import UpcomingBadge from "@/entities/auction/ui/auction-badge/upcoming-badge";
-import { VARIANT_CONFIG } from "@/entities/auction/ui/auction-item-card/model/config";
+import { VARIANT_CONFIG } from "@/entities/auction/ui/auction-item-card/model/constants";
 import { AuctionCardVariantType } from "@/entities/auction/ui/auction-item-card/model/types";
 import AuctionPrice from "@/entities/auction/ui/auction-item-card/ui/auction-price";
 import AuctionTimer from "@/entities/auction/ui/auction-item-card/ui/auction-timer";
